+++ src/sqlalchemy_declarative_extensions/dialects/query.py
@@ -3,13 +3,14 @@ from __future__ import annotations
 from sqlalchemy import func
 from sqlalchemy.engine import Connection
 
-from sqlalchemy_declarative_extensions.dialects import postgresql
+from sqlalchemy_declarative_extensions.dialects import postgresql, snowflake
 from sqlalchemy_declarative_extensions.dialects.mysql.query import (
     check_schema_exists_mysql,
     get_views_mysql,
 )
 from sqlalchemy_declarative_extensions.dialects.postgresql.query import (
     check_schema_exists_postgresql,
+    get_databases_postgresql,
     get_default_grants_postgresql,
     get_functions_postgresql,
     get_grants_postgresql,
@@ -23,6 +24,8 @@ from sqlalchemy_declarative_extensions.dialects.postgresql.query import (
 )
 from sqlalchemy_declarative_extensions.dialects.snowflake.query import (
     check_schema_exists_snowflake,
+    get_databases_snowflake,
+    get_roles_snowflake,
     get_schemas_snowflake,
 )
 from sqlalchemy_declarative_extensions.dialects.sqlite.query import (
@@ -30,6 +33,7 @@ from sqlalchemy_declarative_extensions.dialects.sqlite.query import (
     get_schemas_sqlite,
     get_views_sqlite,
 )
+from sqlalchemy_declarative_extensions.role import Role
 from sqlalchemy_declarative_extensions.sqlalchemy import dialect_dispatch, select
 from sqlalchemy_declarative_extensions.view import View
 
@@ -50,6 +54,11 @@ get_objects = dialect_dispatch(
     postgresql=get_objects_postgresql,
 )
 
+get_databases = dialect_dispatch(
+    postgresql=get_databases_postgresql,
+    snowflake=get_databases_snowflake,
+)
+
 get_default_grants = dialect_dispatch(
     postgresql=get_default_grants_postgresql,
 )
@@ -64,10 +73,13 @@ get_roles = dialect_dispatch(
 
 get_roles = dialect_dispatch(
     postgresql=get_roles_postgresql,
+    snowflake=get_roles_snowflake,
 )
 
 get_role_cls = dialect_dispatch(
     postgresql=lambda _: postgresql.Role,
+    snowflake=lambda _: snowflake.Role,
+    sqlite=lambda _: Role,
 )
 
 get_views = dialect_dispatch(